moved several numbers associated with character creation from playerCharacter.py to characterCreation.py
